RPG fix, modulable indicators and random IA
If the player doesn't click on a tile, it clear the list of tiles possibles and you can click again on the player. Indicators are totally modulable, every method are set in Indicator.py and if you want only one indicator (for example a turret won't move), you only have to redefine indicator_used. Monsters now behave randomly.